Remove tx mutable functions from public scope
The transaction is not expected to be mutated in runtime after
deserialization.

The metadata optimization assumes that, once precomputed, it won't need
to recompute.

If the assumption stands, the code complexity is reduced.
